+++ objectified-ui/public/WHATS_NEW.md
@@ -32,6 +32,8 @@ We continue to improve the platform based on your feedback with improvements and
   - Added ability to apply multiple security schemes to HTTP Operations (and/or)
   - Added ability to specify different security scopes on a per-operation basis
   - Added ability to mark an endpoint as public (i.e. requires no security)
+  - Adds servers section:
+    - Adds ability to create a server, name, and description
   - Configure required OAuth2/OpenID Connect scopes per operation (#421)
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
 - OpenAPI Support: